refactor(addCommand): reduce codebase
