+++ package.json
@@ -14,7 +14,6 @@
     "chalk": "^4.0.0",
     "cli-ux": "^5.4.6",
     "enquirer": "^2.3.5",
-    "pg": "^8.2.1",
     "tslib": "^1"
   },
   "devDependencies": {
@@ -23,7 +22,6 @@
     "@types/chai": "^4",
     "@types/mocha": "^5",
     "@types/node": "^10",
-    "@types/pg": "^7.14.3",
     "chai": "^4",
     "eslint": "^5.13",
     "eslint-config-oclif": "^3.1",
