--- src/functionalities/index.ts
@@ -1,2 +1,0 @@
-export * from "./addNote";
-export * from "./interfaces";
