--- src/functionalities/interfaces.ts
@@ -1,13 +1,0 @@
-export interface ResponseInterface {
-  success: boolean;
-  message: string;
-  note: {
-    title: string;
-    body: string;
-  };
-}
-
-export interface NoteObject {
-  title: string;
-  body: string;
-}
