+++ src/types/addCommand.ts
@@ -1,13 +1,13 @@
-export interface NoteObject {
+export type NoteObject = {
   title: string;
   body: string;
-}
+};
 
-export interface ResponseInterface {
+export type AddNoteResponseObject = {
   success: boolean;
   message: string;
   note: {
     title: string;
     body: string;
   };
-}
+};
